requirements.txt
aiogram>=3.4.1,<4.0.0
python-dotenv>=1.0.1,<2.0.0
streamlit>=1.42.0,<2.0.0
gspread>=6.1.2,<7.0.0
google-auth>=2.38.0,<3.0.0
